backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.3
async-timeout==4.0.3
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.21.3
decli==0.6.1
deprecation==2.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.6
importlib_metadata==7.1.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
postgrest==0.16.2
prompt-toolkit==3.0.36
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
questionary==2.0.1
realtime==1.0.3
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
storage3==0.7.4
StrEnum==0.4.15
supabase-py-async==2.5.6
supafunc==0.4.0
termcolor==2.4.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
uvicorn==0.29.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.18.1
